% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_rasters.R
\name{alignRasters}
\alias{alignRasters}
\title{Align Rasters}
\usage{
alignRasters(referenceRaster = NULL, inputRasters = NULL)
}
\arguments{
\item{referenceRaster}{A \code{SpatRaster} object to be aligned with.}

\item{inputRasters}{A list of \code{SpatRaster} objects to align with the
\code{referenceRaster}.}
}
\value{
A list of \code{SpatRaster} objects that share the same grid and
projection as \code{referenceRaster}.
}
\description{
Aligns multiple rasters with a single reference raster. A raster
will be aligned if it doesn't match the dimensions, resolution, extent,
origin, or CRS projection of the reference raster.
}
\details{
Projecting a raster requires that its cell values be estimated in
accordance with its new projection. Continuous variable rasters will use
bilinear interpolation while categorical (factor) rasters will use
nearest-neighbor sampling.
}
\examples{
\donttest{
#library(TerrainWorksUtils)
#
#referenceRaster <- terra::rast("C:/Work/netmapdata/Puyallup/elev_puy.flt")
#
#inputRasters <- list(
#   gradient = terra::rast("C:/Work/netmapdata/Puyallup/grad_15.tif"),
#   lithology = terra::rast("C:/Work/netmapdata/Puyallup/litho.tif")
# )
#
#alignedRasters <- alignRasters(referenceRaster, inputRasters)
#}
}
\author{
Tate Brasel https://github.com/tabrasel
}
