% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox_to_polygon.R
\name{bbox_to_polygon}
\alias{bbox_to_polygon}
\title{bbox to polygon}
\usage{
bbox_to_polygon(vector)
}
\arguments{
\item{vector}{in sf ou vect format}
}
\value{
sf polygin
}
\description{
extract a bounding box of a vector file and convert
    to a sf polygon
}
\examples{
# bbox_to_polygon(hidro)
}
