% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify_kmeans}
\alias{classify_kmeans}
\title{Classify by kmeans}
\usage{
classify_kmeans(r, num_cluster = 5, decimal_places = 2)
}
\arguments{
\item{r}{SpatRaster raster to be classified}

\item{num_cluster}{number of clusters to be processed}

\item{decimal_places}{define the number of decimal places to legend}
}
\value{
SpatRaster with a categorical raster
}
\description{
Classify a integer or numeric raster using intervals calculated
   by kmeans cluster
}
\examples{
#rk = classify_kmeans(r, 4, decimal_places = 0)
}
