% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correl_feature_elimination.R
\name{correl_feat_elimin}
\alias{correl_feat_elimin}
\title{High correlation feature}
\usage{
correl_feat_elimin(mcor, testcor = c(0.9, 0.925, 0.95, 0.975), graph = TRUE)
}
\arguments{
\item{mcor}{matrix correlation matrix}

\item{testcor}{numeric values of max correlation to be tested}

\item{graph}{boolean plot graphic}
}
\value{
character vector with features names
}
\description{
select variables with correlation is greater then
   a threshold value
}
\examples{
#cor_high = correl_feat_elimin(mcor, testcor = c(0.80, 0.90, 0.95, 0.975))
}
