% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_seeds_rfe.R
\name{create_rfe_seeds}
\alias{create_rfe_seeds}
\title{Create seeds to RFE function}
\usage{
create_rfe_seeds(repeats = 1, number = 5, subsets = (2:10))
}
\arguments{
\item{repeats}{numeric number of repeats}

\item{number}{numeric number of folds}

\item{subsets}{numeric vector of subsets}
}
\value{
list of seeds
}
\description{
Create seeds to RFE function
}
