% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_seeds_train.R
\name{create_train_seeds}
\alias{create_train_seeds}
\title{Create training seeds}
\usage{
create_train_seeds(tuneLength, model = "rf", repeats = 1, number = 10)
}
\arguments{
\item{tuneLength}{numeric number of levels of hyperparameters}

\item{model}{character name of model to be used}

\item{repeats}{numeric number of repeats}

\item{number}{numeric number of CV}
}
\value{
list of seeds
}
\description{
Create caret seeds for training
}
\examples{
seed_train <- create_train_seeds(
  tuneLength = 5, model = "rf",
  repeats = 3, number = 10
)
}
