% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_raster.R
\name{cut_raster}
\alias{cut_raster}
\title{cut raster}
\usage{
cut_raster(r, poly)
}
\arguments{
\item{r}{raster}

\item{poly}{mask polygon}
}
\value{
raster cutted and masked
}
\description{
cut raster by polygon
}
\examples{
#cut_raster <- function(r, poly)
}
