% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dftosf.R
\name{df_to_point}
\alias{df_to_point}
\title{df_to_point - dataframe to point}
\usage{
df_to_point(df, x = "X", y = "Y", crs, todisk = FALSE, pathname = NULL)
}
\arguments{
\item{df}{dataframe data}

\item{x}{character name of column with x coordinates}

\item{y}{character name of column with y coordinates}

\item{crs}{characeter CRS in format "EPSG:XXXXX"}

\item{todisk}{boolean save the sf object do disk}

\item{pathname}{character path + name + extension of file to be saved}
}
\value{
sf object
}
\description{
convert a dataframe with coordinates to a point feature
}
\examples{
df = data.frame(X = c(-42.3,-42.4,-42.5), Y = c(-20.1, -20.2, -20.3))
pt1 = df_to_point(df, x = 'X', y = 'Y', crs = 'EPSG:4326',
               todisk = TRUE, pathname = './pontos/pontos.shp')
}
