% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_grid.R
\name{distance_grid}
\alias{distance_grid}
\title{Distance grid}
\usage{
distance_grid(px, py, nx, ny)
}
\arguments{
\item{px}{numeric vector of x coordinates}

\item{py}{numeric vector of y coordinates}

\item{nx}{numeric number or particions in x direction}

\item{ny}{numeric number of partitions in y direction}
}
\value{
dataframe with point distances to grid
}
\description{
Distance grid
}
