% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epsg_adjust.R
\name{epsg_adjust}
\alias{epsg_adjust}
\title{Adjust epsg}
\usage{
epsg_adjust(epsg, l)
}
\arguments{
\item{l}{vector of character with rasters filenames with path of}
}
\value{

}
\description{
Adiciona informação do epsg para todos os rasters de um vetor
contendo path + nome dos arquivos raster
}
