% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{epsg_br}
\alias{epsg_br}
\title{epsg_br epsg codes to Brazil}
\format{
A data frame with 46 rows and 2 variables:
\describe{
  \item{epsg}{epsg code}
  \item{projecao_datum}{projection and datum of each epsg code}
}
}
\usage{
epsg_br
}
\description{
This dataframe contains a epsg codes most
commonly used in Brazil
}
\keyword{datasets}
