% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_raster_contourn.R
\name{extract_raster_contour}
\alias{extract_raster_contour}
\title{Extract raster contour}
\usage{
extract_raster_contour(r)
}
\arguments{
\item{r}{raster in SpatRaster format}
}
\value{
a polygon with contour of raster
}
\description{
Extract vector of raster contour and create a polygon
}
\examples{
# extract_raster_contour(r)
}
