% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorpho_altrel.R
\name{geomorpho_altrel}
\alias{geomorpho_altrel}
\title{alturas relativas}
\usage{
geomorpho_altrel(saga, mde, pasta, todisk = TRUE)
}
\arguments{
\item{saga}{Rsagamcd object}

\item{mde}{SpatialRaster}

\item{pasta}{character nome da pasta para gravar os arquivos}

\item{todisk}{boolean gravar no disco os rasters gerados}
}
\value{
pilha de rasters (SpatialRaster)
}
\description{
calcula varias métricas de alturas relativas
}
