% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_epsg_SIRGAS}
\alias{get_epsg_SIRGAS}
\title{Get epsg SIRGAS}
\usage{
get_epsg_SIRGAS(lonlat)
}
\arguments{
\item{lonlat}{numeric vector of size 2 with values of longitud and latitud}
}
\value{
numeric epsg number
}
\description{
get the epsg to convert geo to UTM projection and SIRGAS datum
}
\examples{
get_epsg_SIRGAS(c(-42, -20))
}
