% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_epsg_WGS}
\alias{get_epsg_WGS}
\title{Get epsg WGS}
\usage{
get_epsg_WGS(lonlat)
}
\arguments{
\item{lonlat}{numeric vector of size 2 with values of longitud and latitud}
}
\value{
numeric epsg number
}
\description{
get the epsg to convert geo to UTM projection and WGS datum
}
\examples{
get_epsg_WGS(c(-42, -20))
}
