% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{IsBiggerOrEqual}
\alias{IsBiggerOrEqual}
\title{Compare values >=}
\usage{
IsBiggerOrEqual(a, b)
}
\arguments{
\item{a}{first value to be compared}

\item{b}{last value to be compared}
}
\description{
https://stackoverflow.com/questions/9508518/why-are-these-numbers-not-equal
}
\author{
Erdogan CEVHER
}
