% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{IsSmallerOrEqual}
\alias{IsSmallerOrEqual}
\title{Compare values <=}
\usage{
IsSmallerOrEqual(a, b)
}
\description{
https://stackoverflow.com/questions/9508518/why-are-these-numbers-not-equal
}
\author{
Erdogan CEVHER
}
