% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tiles.R
\name{make_tiles}
\alias{make_tiles}
\title{make tiles from a raster}
\usage{
make_tiles(
  r,
  ncol = 2,
  nrow = 2,
  path = NULL,
  sufix = "tile_",
  format = ".tif",
  datatype = "FLT4S"
)
}
\arguments{
\item{r}{SpatRaster}

\item{ncol}{number of vertical divisions}

\item{nrow}{number of horizontal divisions}

\item{path}{folder to save tiles}

\item{sufix}{sufix to be add to filename}

\item{format}{format of file to be saved}

\item{datatype}{type of data (byte, integer, float). Values for datatype are
"INT1U", "INT2U", "INT2S", "INT4U", "atatype = "FLT4S"atatype = "FLT4S"", "FLT4S", "FLT8S". The first
three letters indicate whether the datatype is integer (whole numbers) of
a real number (decimal numbers), the fourth character indicates the number
of bytes used (allowing for large numbers and/or more precision), and the
"S" or "U" indicate whether the values are signed (both negative and
positive) or unsigned (positive values only).}
}
\value{
SpatRaster
}
\description{
split a raster in a number of same size tiles
}
\examples{
# tl = make_tiles(r = r1, ncol = 2, nrow = 2,
#      path = 'tiles', sufix = 'tl', format = '.tif', datatype = "FLT4S")
}
\author{
Cassio Moquedace and Elpidio Filho
}
