% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_splines.R
\name{mpsplinet}
\alias{mpsplinet}
\title{mpsplinet}
\usage{
mpsplinet(
  obj,
  var.name,
  lam = 0.1,
  d = t(c(0, 5, 15, 30, 60, 100, 200)),
  vlow = 0,
  vhigh = 1000,
  show.progress = TRUE
)
}
\arguments{
\item{show.progress}{}
}
\value{

}
\description{
mpsplinet
}
