% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newSummary.R
\name{newSumm_classification}
\alias{newSumm_classification}
\title{New summary classification function}
\usage{
newSumm_classification(data, lev = NULL, model = NULL)
}
\arguments{
\item{data}{data}

\item{lev}{level}

\item{model}{model}
}
\value{
list
}
\description{
summary function with new stats added
}
\examples{
# caret::rfeFuncs$summary = newSummary_classification
}
\author{
Gustavo Veloso
}
