% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newSummary.R
\name{newSummary}
\alias{newSummary}
\title{New summary function}
\usage{
newSummary(data, lev = NULL, model = NULL)
}
\arguments{
\item{data}{data}

\item{lev}{level}

\item{model}{model}
}
\value{

}
\description{
summary function with new stats added
}
\examples{
# caret::rfeFuncs$summary = newSummary
}
\author{
Gustavo Veloso
}
