% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newSummary.R
\name{newSummary_regression}
\alias{newSummary_regression}
\title{New summary regression function}
\usage{
newSummary_regression(data, lev = NULL, model = NULL)
}
\arguments{
\item{data}{data}

\item{lev}{level}

\item{model}{model}
}
\value{
list
}
\description{
summary function with new stats added
}
\examples{
# caret::rfeFuncs$summary = newSummary_regressiion
}
\author{
Gustavo Veloso
}
