% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_info.R
\name{raster_info}
\alias{raster_info}
\title{Recover info from spatRaster
Create dataframe with characteristcs of a list of raster}
\usage{
raster_info(l)
}
\arguments{
\item{l}{vector of rater filenames inclunding path}
}
\value{
dataframe with rasters characteristics.
}
\description{
Create dataframe with characteriscts of a list of raster
}
\examples{
# raster_info(l)

}
