% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_group.R
\name{region_group}
\alias{region_group}
\title{Region Group}
\usage{
region_group(r)
}
\arguments{
\item{r}{SpatRaster}
}
\value{
SpatRaster
}
\description{
associates a unique id for each group of connected pixels of same class
}
\examples{
#
}
