% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_value_raster.R
\name{unique_value_raster}
\alias{unique_value_raster}
\title{Determine raster unique values}
\usage{
unique_value_raster(r)
}
\arguments{
\item{r}{SpatRaster unique layer or  raster stack to be processed}
}
\value{
named character vector with layer names and unique values
}
\description{
Determine raster unique values
}
\examples{
# unique_value_raster(r)
}
