% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_latlong_rasters.R
\name{create_lat_long_rasters}
\alias{create_lat_long_rasters}
\title{Create lat long rasters}
\usage{
create_lat_long_rasters(
  dir_raster = dir_raster,
  extensao = "*.tif",
  dirSaida = dir_raster
)
}
\arguments{
\item{dir_raster}{pasta onde estão os rasters}

\item{extensao}{extensão dos arquivos raster}

\item{dirSaida}{pasta onde será gravado os arquivos de saída}
}
\value{

}
\description{
cria rasters de latitude e longitude
}
