% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_latlong_rasters.R
\name{create_lat_long_rasters}
\alias{create_lat_long_rasters}
\title{Create lat long rasters}
\usage{
create_lat_long_rasters(
  rr,
  todisk = TRUE,
  extensao = ".tif",
  dir_saida = getwd()
)
}
\arguments{
\item{rr}{raster de referencia}

\item{todisk}{boolean gravar ou não os arquivos no disco}

\item{extensao}{extensão dos arquivos raster .tif, .asc...}

\item{dirSaida}{pasta onde será gravado os arquivos de saída}
}
\value{
lat e long em uma lista
}
\description{
cria rasters de latitude e longitude
}
\author{
Elpidio Filho
}
