% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epsg_adjust.R
\name{epsg_adjust}
\alias{epsg_adjust}
\title{Adjust Raster epsg}
\usage{
epsg_adjust(l, epsg)
}
\arguments{
\item{l}{character filename with path of rasters files}

\item{epsg}{character epsg in format: 'epsg:xxxxx'. Ex.'epsg:32723'}
}
\value{
SpatRaster
}
\description{
Add epsg information for all rasters of a vector
    containing path + raster filenames
}
\examples{
l = c("./rasters/mde.tif', './rasters/slope.tif")
epsg_adjust(l, "epsg:32723")
}
