% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euc_distance.R
\name{euc_dist}
\alias{euc_dist}
\title{euc_distance - Euclidean distance to a vector}
\usage{
euc_dist(r, vct, todisk = FALSE, path = NULL, filename = NULL, ext = ".tif")
}
\arguments{
\item{r}{SpatRaster raster}

\item{vct}{SpatVector point, line or polygon}

\item{todisk}{Boolean if true save to disk}

\item{filename}{character raster filename}

\item{ext}{character file extension  (.tif, .asc, .img)}

\item{folder}{character path to save}
}
\value{
SpatRaster
}
\description{
Calculates euclidean distance to point, line or polygon
}
\examples{
# rdist = euc_dist(r, vec_pt, todisk = FALSE)
}
