% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorpho_altrel.R
\name{geomorpho_altrel}
\alias{geomorpho_altrel}
\title{relative height metrics}
\usage{
geomorpho_altrel(saga, mde, todisk = FALSE, folder = NULL)
}
\arguments{
\item{saga}{Rsagamcd object}

\item{mde}{SpatialRaster raster}

\item{todisk}{logical write the generated raster to disk}

\item{folder}{character folder name to save files}
}
\value{
SpatialRaster with 5 layers
}
\description{
calculates various relative height metrics
    Slope_Height, Valley_Depth, Normalized_Height,
    Standardized_Height, Mid_Slope_Positon using SAGA GIS
}
\examples{
# library(rsagacmd)
# saga <- saga_gis()
# frel <- geomorpho_altrel (saga, mde = dem, folder = './morpho',
#                          todisk = TRUE)
}
