% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorpho_altrel.R
\name{geomorpho_altrel}
\alias{geomorpho_altrel}
\title{relative height metrics}
\usage{
geomorpho_altrel(saga, mde, pasta, todisk = TRUE)
}
\arguments{
\item{saga}{Rsagamcd object}

\item{mde}{SpatialRaster raster}

\item{pasta}{character folder name to save files}

\item{todisk}{boolean write the generated rasters to disk}
}
\value{
raster stack (SpatialRaster)
}
\description{
calculates various relative height metrics
    Slope_Height, Valley_Depth, Normalized_Height,
    Standardized_Height, Mid_Slope_Positon
}
