% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorpho_altrel.R
\name{list_to_fileraster}
\alias{list_to_fileraster}
\title{Title convert lista de rasters em uma pilha de rasters}
\usage{
list_to_fileraster(terrain, path)
}
\arguments{
\item{path}{character folder}

\item{listrrain}{list or SpatRaster}
}
\value{
SpatRaster
}
\description{
Title convert lista de rasters em uma pilha de rasters
}
\examples{
# rl = list_to_fileraster(alrel, "morpho")
}
