% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tiles.R
\name{make_tiles}
\alias{make_tiles}
\title{make tiles from a raster}
\usage{
make_tiles(
  r,
  ncol = 2,
  nrow = 2,
  path = "./tiles/",
  sufix = "tile_",
  format = ".tif"
)
}
\arguments{
\item{r}{SpatRaster}

\item{ncol}{número de divisões verticais}

\item{nrow}{número de divisões horizontais}

\item{path}{pasta onde será salvo os tiles (com / no final)}

\item{sufix}{sufixo a ser adicionada ao nome dos arquivos}

\item{format}{formato do arquivo a ser gravado (com ponto no inicio)}
}
\value{
SpatRaster
}
\description{
make tiles from a raster
}
\examples{
tl = make_tiles(r = r1, ncol = 2, nrow = 2,
     path = './tiles', sufix = 'tl', format = '.tif')
}
\author{
Cassio Moquedace e Elpidio Filho
}
