% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tiles.R
\name{make_tiles}
\alias{make_tiles}
\title{make tiles from a raster}
\usage{
make_tiles(
  r,
  ncol = 2,
  nrow = 2,
  path = NULL,
  sufix = "tile_",
  format = ".tif"
)
}
\arguments{
\item{r}{SpatRaster}

\item{ncol}{number of vertical divisions}

\item{nrow}{number of horizontal divisions}

\item{path}{folder to save tiles}

\item{sufix}{sufix to be add to filename}

\item{format}{format of file to be saved}
}
\value{
SpatRaster
}
\description{
spliy a raster in a number of same size tiles
}
\examples{
# tl = make_tiles(r = r1, ncol = 2, nrow = 2,
#      path = 'tiles', sufix = 'tl', format = '.tif')
}
\author{
Cassio Moquedace and Elpidio Filho
}
