% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moda.R
\name{mode_freq}
\alias{mode_freq}
\title{Frequency of modal values}
\usage{
mode_freq(r, ncores = 4)
}
\arguments{
\item{r}{SpatRaster}

\item{ncores}{integer number of cores for multiprocessing}
}
\value{
SpatRaster
}
\description{
Frequency of modal values
}
\author{
<elpidio@ufv.br>
}
