% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_dummy.R
\name{raster_dummy}
\alias{raster_dummy}
\title{Converts Categorical Raster to Dummy Raster}
\usage{
raster_dummy(r)
}
\arguments{
\item{r}{raster with categorical data}
}
\value{
SpatRaster with n layers one for each factor level
}
\description{
Convert raster file containing n classes
    in n files containing one class
    type presence (1) and absence (0)
    Output a multiple raster (SpatRaster) from the terra packet
}
\examples{
#dumr = raster_dummy(r)
}
